% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lkm.R
\name{add_lkm}
\alias{add_lkm}
\title{Construction function to add unit km}
\usage{
add_lkm(x)
}
\arguments{
\item{x}{Object with class "data.frame", "matrix",  "numeric" or "integer"}
}
\value{
Objects of class "data.frame" or "units"
}
\description{
\code{add_lkm} just  add unit 'km' to different R objects
}
\examples{
\dontrun{
a <- add_lkm(rnorm(100)*10)
plot(a)
b <- add_lkm(matrix(rnorm(100)*10, ncol = 10))
print(head(b))
}
}
\seealso{
Other Add distance unitts:
\code{\link[=add_miles]{add_miles()}}
}
\concept{Add distance unitts}
