% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polid.R
\name{add_polid}
\alias{add_polid}
\title{Add polygon id to lines road network}
\usage{
add_polid(polyg, street, by)
}
\arguments{
\item{polyg}{sf object POLYGON
it is transformed to "sf" with emissions.}

\item{street}{streets road network class 'sf'}

\item{by}{Character indicating the column woith the id in polyg}
}
\description{
Sometimes you need to add polygon id into your streets road network.
\code{\link{add_polid}} add add_polid id into your road network cropping your
network by
}
\examples{
\dontrun{
#to do
}
}
\seealso{
\code{\link{emis_to_streets}}
}
