% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt}
\alias{adt}
\title{Average daily traffic (ADT) from hourly traffic data.}
\usage{
adt(
  pc,
  lcv,
  hgv,
  bus,
  mc,
  p_pc,
  p_lcv,
  p_hgv,
  p_bus,
  p_mc,
  feq_pc = 1,
  feq_lcv = 1.5,
  feq_hgv = 2,
  feq_bus = 2,
  feq_mc = 0.5
)
}
\arguments{
\item{pc}{numeric vector for passenger cars}

\item{lcv}{numeric vector for light commercial vehicles}

\item{hgv}{numeric vector for heavy good vehicles or trucks}

\item{bus}{numeric vector for bus}

\item{mc}{numeric vector for motorcycles}

\item{p_pc}{data-frame profile for passenger cars, 24 hours only.}

\item{p_lcv}{data-frame profile for light commercial vehicles, 24 hours only.}

\item{p_hgv}{data-frame profile for heavy good vehicles or trucks, 24 hours only.}

\item{p_bus}{data-frame profile for bus, 24 hours only.}

\item{p_mc}{data-frame profile for motorcycles, 24 hours only.}

\item{feq_pc}{Numeric, factor equivalence}

\item{feq_lcv}{Numeric, factor equivalence}

\item{feq_hgv}{Numeric, factor equivalence}

\item{feq_bus}{Numeric, factor equivalence}

\item{feq_mc}{Numeric, factor equivalence}
}
\value{
numeric vector of total volume of traffic per link as ADT
}
\description{
\code{\link{adt}} calculates ADT based on hourly traffic data.
}
\examples{
{
data(net)
data(pc_profile)
p1 <- pc_profile[, 1]
adt1 <- adt(pc = net$ldv*0.75,
            lcv = net$ldv*0.1,
            hgv = net$hdv,
            bus = net$hdv*0.1,
            mc = net$ldv*0.15,
            p_pc = p1,
            p_lcv = p1,
            p_hgv = p1,
            p_bus = p1,
            p_mc = p1)
head(adt1)
}
}
