% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.EmissionFactorsList.R
\name{as.EmissionFactorsList}
\alias{as.EmissionFactorsList}
\alias{EmissionFactorsList}
\title{as.EmissionFactorsList}
\usage{
EmissionFactorsList(ef, ...)
}
\arguments{
\item{ef}{Object with class "EmissionFactors"}

\item{...}{ignored}

\item{lfx}{Logical value to determine if the returning object will be
a list of functions or not. Each function is dependent on the speed.}

\item{mass}{Character to determine the unit of the mass. Default is "g"}

\item{distance}{Character to determine the distance unit. Default is "km"}
}
\description{
Returns a tranformed object with class "EmissionFactors" and
units g/km or "EmissionsFactorsList". This functions has arguments to change
the units. The type of objects supported are of classes "matrix",
"data.frame", "list" and "numeric". If the class of the object is "matrix"
this function returns a dataframe. When then object is a list with numerics
or a list with functions it returns an "EmissionFactorsList".
}
\details{
Construction function for class "EmissionFactorsList"
}
