% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.Vehicles.R
\name{as.Vehicles}
\alias{as.Vehicles}
\title{Construction function for class "Vehicles"}
\usage{
as.Vehicles(veh, ...)
}
\arguments{
\item{veh}{Object with class "Vehicles"}
}
\value{
Objects of class "Vehicles" or "units"
}
\description{
Returns a tranformed object with class "Vehicles" and units
 1/h. The type of objects supported are of classes "matrix", "data.frame",
 "numeric" and "array". If the object is a matrix it is converted to data.frame.
 If the object is "numeric" it is converted to class "units". The function
 \code{\link{emis_paved}} needs veh to be an array, therefore in this case,
 veh must be an array in the total fleet at each street and dimensions
 total fleet, hours and days
}
\examples{
\dontrun{
data(net)
lt <- as.Vehicles(net$hdv)
class(lt)
plot(lt)
}
}
