% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{check_nt}
\alias{check_nt}
\title{Check the max number of threads}
\usage{
check_nt()
}
\value{
Integer with the max number of threads
}
\description{
\code{get_threads} check the number of threads in this machine
}
\examples{
{
  check_nt()
}
}
