% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\docType{data}
\name{decoder}
\alias{decoder}
\title{Description data.frame for MOVES}
\format{
A data frame with 69 rows and 4 columns:
\describe{
  \item{CategoryField}{dayID, sourceTypID, roadTypeID, pollutantID and procesID}
  \item{pollutantID}{Associated number}
  \item{Description}{Associatd description}
  \item{V4}{pollutants}
}
}
\source{
US/EPA MOVES
}
\usage{
data(decoder)
}
\description{
A data.frame descriptors to use MOVES functions
}
\keyword{datasets}
