% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_china.R
\name{ef_china_long}
\alias{ef_china_long}
\title{Chinese emission factors by emissions standard}
\usage{
ef_china_long(v = "PV", t = "Small", f = "G", standard, p)

ef_china_long(v = "PV", t = "Small", f = "G", standard, p)
}
\arguments{
\item{v}{Character; category vehicle: "PV" for Passenger Vehicles or 'Trucks"}

\item{t}{Character; sub-category of of vehicle: PV Gasoline:  "Mini", "Small","Medium",
"Large", "Taxi", "Motorcycles", "Moped", PV Diesel: "Mediumbus", "Largebus",
 "3-Wheel". Trucks: "Mini", "Light" , "Medium", "Heavy"}

\item{f}{Character;fuel: "G", "D", "CNG", "ALL"}

\item{standard}{Character vector; "PRE", "I", "II", "III", "IV", "V".}

\item{p}{Character; pollutant: "CO", "NOx","HC", "PM", "Evaporative_driving"
or "Evaporative_parking"}
}
\value{
long data.frame

long data.frame
}
\description{
Chinese emission factors in long format

Correction of Chinese emission
}
\examples{
{
\dontrun{
# Do not run
}
}
{
ef_china_long(standard = "I", p = "CO")
}
}
\seealso{
Other China:
\code{\link[=ef_china]{ef_china()}},
\code{\link[=ef_china_det]{ef_china_det()}},
\code{\link[=ef_china_h]{ef_china_h()}},
\code{\link[=ef_china_hu]{ef_china_hu()}},
\code{\link[=ef_china_s]{ef_china_s()}},
\code{\link[=ef_china_speed]{ef_china_speed()}},
\code{\link[=ef_china_te]{ef_china_te()}},
\code{\link[=ef_china_th]{ef_china_th()}},
\code{\link[=emis_china]{emis_china()}},
\code{\link[=emis_long]{emis_long()}}
}
\concept{China}
