% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Emissions estimation hourly for the of the week}
\usage{
emis(veh, lkm, ef, speed, agemax, profile, hour = 1, day = 1, array = F)
}
\arguments{
\item{veh}{Numeric vector with length of elements equals to number of streets}

\item{lkm}{Length of each link}

\item{ef}{List of functions of emission factors}

\item{speed}{List of speeds}

\item{agemax}{Age of oldest vehicles for that category}

\item{profile}{Numerical or dataframe with nrows equal to 24 and ncol 7 day of the week}

\item{hour}{Number of considered hours in estimation}

\item{day}{Number of considered days in estimation}

\item{array}{When FALSE produces a dataframe of the estimation. When TRUE expects a
profile as a dataframe producing an array with dimensions (streets x columns x hours x days)}
}
\value{
emission estimation  g/h
}
\description{
The vehicular emissions are estimated as the product of the vehicles on a
road, length of the road, emission factor avaliated at the respective speed.
 \eqn{E = VEH*LENGTH*EF(speed)}
}
\examples{
\dontrun{
# Do not run
pc <- rnorm(n = 100, mean = 500, sd = 100)
lkm <-  rnorm(n = 100, mean = 10, sd = 2)
ef <- ef_ldv_speed(v = "PC",t = "PRE_ECE", cc = "ALL", f = "G",
                    eu = "PRE", p = "CO")

emi  <- emis(veh = pc, lkm = lkm, ef = ef, speed = 30 )
}
}
