% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_chem.R
\name{emis_chem}
\alias{emis_chem}
\title{Aggregate emissions}
\usage{
emis_chem(dfe, mechanism, colby)
}
\arguments{
\item{dfe}{data.frame with column `emissions` in grams and `pollutant`.}

\item{mechanism}{Character, any of "SAPRC", "RACM", "RADM2", "CBMZ",
"MOZART", "SAPRC99", "CB05", "CB06CMAQ", "RACM2CMAQ", "SAPRC99CMAQ",
"SAPRC07CMAQ" or "SAPRC07A".}

\item{colby}{Vector for aggregating extra column. For instance, region or
province}
}
\value{
data.frame with lumped groups by chemical mechanism. It transform
emissions in grams to mol.
}
\description{
\code{\link{emis_chem}} aggregates emissions by chemical mechanism
and convert grams to mol. This function reads all hydrocarbos and respective
criteria polluants specified in \code{\link{ef_ldv_speed}} and \code{\link{ef_hdv_speed}}.
}
\examples{
{
# CO
df <- data.frame(emission = 1:10)
df$pollutant = "CO"
df$emission <- units::set_units(df$emission, "g")
emis_chem(df, "CBMZ")
# hexanal
df$pollutant = "hexanal"
emis_chem(df, "CBMZ")
# propadiene and NO2
df2 <- df1 <- df
df1$pollutant = "propadiene"
df2$pollutant = "NO2"
emis_chem(rbind(df1, df2), "RADM2")
}
}
\seealso{
\code{\link{ef_ldv_speed}} \code{\link{ef_hdv_speed}} \code{\link{speciate}}
}
