% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_cold.R
\name{emis_cold}
\alias{emis_cold}
\title{Estimation of cold start emissions hourly for the of the week}
\usage{
emis_cold(veh, lkm, ef, efcold, beta, speed, agemax, profile, hour = 1,
  day = 1, array = F)
}
\arguments{
\item{veh}{Numeric vector with length of elements equals to number of streets}

\item{lkm}{Length of each link}

\item{ef}{List of functions of emission factors}

\item{efcold}{List of functions of cold start emission factors}

\item{beta}{Datraframe with the hourly cold-start distribution to each day
of the period. Number of rows are hours and columns are days}

\item{speed}{List of speeds}

\item{agemax}{Age of oldest vehicles for that category}

\item{profile}{Numerical or dataframe with nrows equal to 24 and ncol 7 day
of the week}

\item{hour}{Number of considered hours in estimation}

\item{day}{Number of considered days in estimation}

\item{array}{When FALSE produces a dataframe of the estimation. When TRUE
expects a profile as a dataframe producing an array with dimensions
(streets x columns x hours x days)}
}
\value{
emission estimation  g/h
}
\description{
The vehicular emissions are estimated as the product of the vehicles on a
road, length of the road, emission factor avaliated at the respective speed.
The estimation considers beta
parameter, the fraction of mileage driven
}
\note{
Actually dcold is not necessary, it would be enough to multiply
an existing cold-start distribution with the daily profile, but it was added
because it is important to clarify both, the data and the concepts
}
\examples{
\dontrun{
# Do not run
}
}
