% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_cold_td.R
\name{emis_cold_td}
\alias{emis_cold_td}
\title{Estimation of cold start emissions with top-down approach}
\usage{
emis_cold_td(veh, lkm, ef, efcold, beta, pro_month, verbose = FALSE)
}
\arguments{
\item{veh}{"Vehicles" data-frame or spatial feature, wwhere columns are the
 age distribution of that vehicle. and rows each simple feature or region.
The number of rows is equal to the number of streets link}

\item{lkm}{Numeric; mileage by the age of use of each vehicle.}

\item{ef}{Numeric; emission factor with}

\item{efcold}{Data.frame or list of data.frames. When it is a data.frame, each column is for each
type of vehicle by age of use, rows are are each simple feature. When it is
a list of data.frames, each element of the list it is a month of the year.
efcold should be different for each simple feature, assuming different average temperature.}

\item{beta}{Data.frame with the fraction of cold starts. The rows are the fraction
for each spatial feature or subregion, the columns are the age of use of vehicle.}

\item{pro_month}{Numeric; montly profile to distribuite annual mileage in each month.}

\item{verbose}{Logical; To show more information}
}
\value{
Emissions data.frame
}
\description{
\code{emis_cold_td} estimates cld start emissions with
a top-down appraoch. This is, annual or monthly emissions or region.
Especifically, the emissions are esitmated for row of the simple feature (row
of the spatial feature).

In general was designed so that each simple feature is a region with
different average monthly temperature.
This funcion, as other in this package, adapts to the class of the input data.
providing flexibility to the user.
}
\examples{
\dontrun{
# Do not run
}
}
