% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_hot_td.R
\name{emis_hot_td}
\alias{emis_hot_td}
\title{Estimation of emissions Hot Exhaust top down}
\usage{
emis_hot_td(veh, lkm, ef, speed = 34,
  agemax = ifelse(is.data.frame(veh), ncol(veh), ncol(veh[[1]])),
  verbose = FALSE)
}
\arguments{
\item{veh}{"Vehicles" data-frame or list of "Vehicles" data-frame. Each data-frame
as number of columns matching the age distribution of that ype of vehicle.
The number of rows is equal to the number of streets link}

\item{lkm}{Length of each link}

\item{ef}{List of functions of emission factors}

\item{speed}{Speed data-frame with number of columns as hours}

\item{agemax}{Age of oldest vehicles for that category}

\item{verbose}{Logical; To show more information}
}
\value{
emission estimation  g/h
}
\description{
\code{\link{emis}} estimates vehicular emissions as the product of the
vehicles on a road, length of the road, emission factor avaliated at the
respective speed. \eqn{E = VEH*LENGTH*EF(speed)}
}
\note{
If the user apply a top-down approach, the resulting units will be
according its own data. For instance, if the vehicles are veh/day, the units
of the emissions implicitly will be g/day.
Hour and day will be deprecate because they can be infered from the profile
matrix.
}
\examples{
\dontrun{
# Do not run
}
}
