% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_merge.R
\name{emis_merge}
\alias{emis_merge}
\title{Merge several emissions files returning data-frames or 'sf' of lines}
\usage{
emis_merge(path = "emi", pol = "CO", what = "STREETS.rds", net,
  streets = T)
}
\arguments{
\item{path}{Character. Path where emissions are located}

\item{pol}{Character. Pollutant.}

\item{what}{Character. Word to search the emissions names, "STREETS", "DF" or
whatever name. It is important to include the extension .'rds'}

\item{net}{'Spatial feature' or 'SpatialLinesDataFrame' with the streets.
It is expected #' that the number of rows is equal to the number of rows of
street emissions. If #' not, the function will stop.}

\item{streets}{Logical. If true, \code{\link{emis_merge}} will read the street
emissions created with \code{\link{emis_post}} by "streets_wide", returning an
object with class 'sf'. If false, it will read the emissions data-frame and
rbind them.}
}
\value{
'Spatial feature' of lines or a dataframe of emissions
}
\description{
\code{\link{emis_merge}} reads rds files and returns a data-frame
 or an object of 'spatial feature' of streets, merging several files.
}
\examples{
\dontrun{
# Do not run

}
}
