% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_source.R
\name{emis_source}
\alias{emis_source}
\title{A function to source vein scripts}
\usage{
emis_source(path = "est", pattern = ".R", ignore = "~", first,
  ask = TRUE, recursive = TRUE, full.names = TRUE)
}
\arguments{
\item{path}{Character; path to source scripts. Default is "est".}

\item{pattern}{Character; extensions of R scripts. Default is ".R".}

\item{ignore}{Character; caracter to be excluded. Default is "~". Sometimes,
the OS creates automatic back-ups, for instance "run.R~", the ideia is to
avoid sourcing these files.}

\item{first}{Character; first script.}

\item{ask}{Logical; Check inputs or not. Default is "FALSE". It allows to
stop inputs}

\item{recursive}{Logical; recursive or not. Default is "TRUE"}

\item{full.names}{Logical; full.names or not. Default is "TRUE".}
}
\description{
\code{\link{emis_source}} source vein scripts
}
\examples{
\dontrun{
# Do not run
}
}
