% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_to_streets.R
\name{emis_to_streets}
\alias{emis_to_streets}
\title{Emis to streets distribute top-down emissions into streets}
\usage{
emis_to_streets(streets, dfemis, by = "ID", verbose = TRUE)
}
\arguments{
\item{streets}{sf object with geometry 'LINESTRING' or 'MULTILINESTRING'. Or
SpatialLinesDataFrame}

\item{dfemis}{data.frame with emissions}

\item{by}{Character indicating the columns that must be present in both
'street' and 'dfemis'}

\item{verbose}{Logical; to show more info.}
}
\description{
\code{\link{emis_to_streets}} allocates emissions proportionally to
each feature. "Spatial" objects are converter to "sf" objects. Currently,
'LINESTRING' or 'MULTILINESTRING' supported. The emissions are distributed
in each street.
}
\note{
When spobj is a 'Spatial' object (class of sp), they are converted
 into 'sf'.
}
\examples{
\dontrun{
# TO DO
}
}
\seealso{
\code{\link{add_polid}}
}
