% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactors.default.R
\name{EmissionFactors.default}
\alias{EmissionFactors.default}
\title{EmissionFactors}
\usage{
\method{EmissionFactors}{default}(ef, mass = "g", distance = "km", ...)
}
\arguments{
\item{ef}{Object with class "EmissionFactors"}

\item{mass}{Character to determine the unit of the mass. Default is "g"}

\item{distance}{Character to determine the distance unit. Default is "km"}

\item{...}{ignored}
}
\value{
Objects of class "EmissionFactors" or "units"
}
\description{
Returns a tranformed object with class "EmissionFactors" and
units g/km. This functions has arguments to change
the units.
}
\details{
Construction function for class "EmissionFactors"
}
\note{
If the class ob the object is functions, as.EmissionFactors won't
append another class
}
\examples{
\dontrun{
data(fe2015)
names(fe2015)
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- as.EmissionFactors(df)
class(ef1)
head(ef1)
}
}
