% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactors.R
\name{EmissionFactors}
\alias{EmissionFactors}
\title{EmissionFactors}
\usage{
EmissionFactors(ef, ...)
}
\arguments{
\item{ef}{Object which can be "matrix", "data.frame" or "numeric"}

\item{...}{ignored}
}
\value{
Objects of class "EmissionFactors" or "units"
}
\description{
Generic function for class "EmissionFactors"
}
\details{
Generic function for lass "EmissionFactors"
}
