% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactorsList.R
\name{EmissionFactorsList}
\alias{EmissionFactorsList}
\title{Generic function EmissionFactorsList}
\usage{
EmissionFactorsList(ef, ...)
}
\arguments{
\item{ef}{object}

\item{...}{ignored}
}
\description{
Generic function
}
