% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{emissions}
\alias{emissions}
\title{Emissions estimation hourly and daily}
\usage{
emissions(veh, lkm, ef, speed, agemax = 40, day = F, k = 1)
}
\arguments{
\item{veh}{Numeric vector with length of elements equals to number of streets}

\item{lkm}{Length of each link}

\item{ef}{Emission factor function}

\item{speed}{Speed. It can be a numeric vector or a list.}

\item{agemax}{Age of oldest vehicles for that category}

\item{day}{When false returns a numeric factor with one hour emissions, when
TRUE expects emission factor as a list with number of elements equals to the
number of respective vehicular category and speed as list with length equals
to 24, one per hour, returning an array}

\item{k}{Factor to extrapolate the vehicular flux to other hours}
}
\value{
emissions estimation g/h
}
\description{
The vehicular emissions are estimated as the product of the vehicles on a
road, length of the road (units depend on the emission factor), emission
factor avaliated at the respective speed. \eqn{E = VEH*LENGTH*EF(speed)}
}
\examples{
 \dontrun{
# Do not run
pc <- rnorm(n = 100, mean = 500, sd = 100)
lkm <-  rnorm(n = 100, mean = 10, sd = 2)
ef <- ef_ldv_speed(v = "PC",t = "PRE_ECE", cc = "ALL", f = "G",
                    eu = "PRE", p = "CO")

emi  <- emissions(veh = pc, lkm = lkm, ef = ef, speed = 30 )
}
}
