% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_24.R
\name{emissions_24}
\alias{emissions_24}
\title{Emissions estimation hourly for the of the week}
\usage{
emissions_24(veh, lkm, ef, speed)
}
\arguments{
\item{veh}{Numeric vector with length of elements equals to number of streets}

\item{lkm}{Length of each link}

\item{ef}{List of functions of emission factors}

\item{speed}{List of speeds}
}
\value{
emission estimation g/h
}
\description{
The vehicular emissions are estimated as the product of the vehicles on a
road, length of the road (units depend on the emission factor), emission
factor avaliated at the respective speed. \eqn{E = VEH*LENGTH*EF(speed)}
}
\examples{
\dontrun{
# Do not run
pc <- rnorm(n = 100, mean = 500, sd = 100)
lkm <-  rnorm(n = 100, mean = 10, sd = 2)
ef <- ef_ldv_speed(v = "PC",t = "PRE_ECE", cc = "ALL", f = "G",
                    eu = "PRE", p = "CO")

emi  <- emissions_24(veh = pc, lkm = lkm, ef = ef, speed = 30 )
}
}
