% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsArray.R
\name{EmissionsArray}
\alias{EmissionsArray}
\title{EmissionsArray}
\usage{
EmissionsArray(e, ...)
}
\arguments{
\item{e}{Object which can be "array"}

\item{...}{ignored}
}
\value{
Objects of class "EmissionsArray"
}
\description{
Generic function for class "EmissionsArray"
}
\details{
Generic function for lass "EmissionsArray"
}
