% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsList.R
\name{EmissionsList}
\alias{EmissionsList}
\title{EmissionsList}
\usage{
EmissionsList(e, ...)
}
\arguments{
\item{e}{Object which can be "list"}

\item{...}{ignored}
}
\value{
Objects of class "Emissionslist"
}
\description{
Generic function for class "EmissionsList"
}
\details{
Generic function for lass "EmissionsList"
}
