% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ef_ref.R
\name{get_ef_ref}
\alias{get_ef_ref}
\title{Get ef reference data}
\usage{
get_ef_ref(ref)
}
\arguments{
\item{ref}{Character; The ef model required (e.g. "eea" for ef_eea)}
}
\description{
Get the reference data used to build the emission factor (ef)
model applied by vein.
}
\note{
This function is a shortcut to access unexported ef model information
in vein.
}
\examples{
\dontrun{
get_ef_ref("eea")
}
}
