% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GriddedEmissionsArray.R
\name{GriddedEmissionsArray}
\alias{GriddedEmissionsArray}
\alias{print.GriddedEmissionsArray}
\alias{summary.GriddedEmissionsArray}
\alias{plot.GriddedEmissionsArray}
\title{Construction function for class "GriddedEmissionsArray"}
\usage{
GriddedEmissionsArray(x, ..., cols, rows, times = ncol(x), rotate = "default")

\method{print}{GriddedEmissionsArray}(x, ...)

\method{summary}{GriddedEmissionsArray}(object, ...)

\method{plot}{GriddedEmissionsArray}(x, ..., times = 1)
}
\arguments{
\item{x}{Object with class "SpatialPolygonDataFrame", "sf" "data.frame" or
"matrix"}

\item{...}{ignored}

\item{cols}{Number of columns}

\item{rows}{Number of rows}

\item{times}{Number of times}

\item{rotate}{Character, rotate array:"default",  "left", "right",
"cols","rows", "both", "br", "colsbr", "rowsbr", "bothbr".
br means starting a matrix byrow}

\item{object}{object with class "EmissionsArray'}
}
\value{
Objects of class "GriddedEmissionsArray"
}
\description{
\code{GriddedEmissionsArray} returns a transformed object with class
"EmissionsArray" with 4 dimensions.
}
\examples{
\dontrun{
data(net)
data(pc_profile)
data(fe2015)
data(fkm)
PC_G <- c(33491,22340,24818,31808,46458,28574,24856,28972,37818,49050,87923,
          133833,138441,142682,171029,151048,115228,98664,126444,101027,
          84771,55864,36306,21079,20138,17439, 7854,2215,656,1262,476,512,
          1181, 4991, 3711, 5653, 7039, 5839, 4257,3824, 3068)
veh <- data.frame(PC_G = PC_G)
pc1 <- my_age(x = net$ldv, y = PC_G, name = "PC")
pcw <- temp_fact(net$ldv+net$hdv, pc_profile)
speed <- netspeed(pcw, net$ps, net$ffs, net$capacity, net$lkm, alpha = 1)
pckm <- units::set_units(fkm[[1]](1:24), "km")
pckma <- cumsum(pckm)
cod1 <- emis_det(po = "CO", cc = 1000, eu = "III", km = pckma[1:11])
cod2 <- emis_det(po = "CO", cc = 1000, eu = "I", km = pckma[12:24])
#vehicles newer than pre-euro
co1 <- fe2015[fe2015$Pollutant=="CO", ] #24 obs!!!
cod <- c(co1$PC_G[1:24]*c(cod1,cod2),co1$PC_G[25:nrow(co1)])
lef <- ef_ldv_scaled(co1, cod, v = "PC", t = "4S", cc = "<=1400",
                     f = "G",p = "CO", eu=co1$Euro_LDV)
E_CO <- emis(veh = pc1,lkm = net$lkm, ef = lef, speed = speed, agemax = 41,
              profile = pc_profile, simplify = TRUE)
class(E_CO)
E_CO_STREETS <- emis_post(arra = E_CO, pollutant = "CO", by = "streets",
                          net = net, k = units::set_units(1, "1/h"))
g <- make_grid(net, 1/102.47/2, 1/102.47/2) #500m in degrees
E_CO_g <- emis_grid(spobj = E_CO_STREETS, g = g, sr= 31983)
plot(E_CO_g["V9"])
# check all
rots <- c("default", "left", "right",
          "cols","rows", "both",
          "br", "colsbr", "rowsbr", "bothbr")
oldpar <- par()
par(mfrow = c(2,5))
lg <- lapply(seq_along(rots), function(i){
            x <- GriddedEmissionsArray(E_CO_g,
                                rows = 19,
                                cols = 23,
                                times = 168,
                                rotate = rots[i])
         plot(x, main = rots[i])
        })

par(mfrow = c(1,1))
}
}
