% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invcop.R
\name{invcop}
\alias{invcop}
\title{Helper function to copy and zip projects}
\usage{
invcop(in_name = getwd(), out_name, all = FALSE, main = TRUE, ef = TRUE,
  est = TRUE, network = TRUE, veh_rds = FALSE, veh_csv = TRUE,
  zip = TRUE)
}
\arguments{
\item{in_name}{Character; Name of current project.}

\item{out_name}{Character; Name of outtput project.}

\item{all}{Logical; copy ALL (and for once) or not.}

\item{main}{Logical; copy or not.}

\item{ef}{Logical; copy or not.}

\item{est}{Logical; copy or not.}

\item{network}{Logical; copy or not.}

\item{veh_rds}{Logical; copy or not.}

\item{veh_csv}{Logical; copy or not.}

\item{zip}{Logical; zip or not.}
}
\value{
emission estimation  g/h
}
\description{
\code{invcop} help to copy and zip projects
}
\note{
This function was created to copy and zip project without the emis.
}
\examples{
\dontrun{
# Do not run
}
}
