% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_wide.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Transform data.frame from long to wide format}
\usage{
long_to_wide(df, column_with_new_names = names(df)[1],
  column_with_data = "emission", column_fixed, geometry)
}
\arguments{
\item{df}{data.frame with three column. Forst has the new names, secon}

\item{column_with_new_names}{Character, column that has new column names}

\item{column_with_data}{Character column with data}

\item{column_fixed}{Character, optional, column that will remain fixed}
}
\value{
wide data.frame.
}
\description{
\code{\link{lon}} aggregates emissions by chemical mechanism
and convert grams to mol. This function reads all hydrocarbos and respective
criteria polluants specified in \code{\link{ef_ldv_speed}} and \code{\link{ef_hdv_speed}}.
}
\examples{
{
df <- data.frame(pollutant = rep(c("CO", "propadiene", "NO2"), 10),
emission = vein::Emissions(1:30),
region = rep(letters[1:2], 15))
df
long_to_wide(df)
long_to_wide(df, column_fixed = "region")
}
}
\seealso{
\code{\link{emis_hot_td}}  \code{\link{emis_cold_td}}
}
