% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matvect.R
\name{matvect}
\alias{matvect}
\title{Matrix and vector multiplication}
\usage{
matvect(df, x, by = "row")
}
\arguments{
\item{df}{Numeric Data-frame or matrix.}

\item{x}{Numeric vector.}

\item{by}{Character, with two value "row" or "col"}
}
\value{
data-frame
}
\description{
\code{matvect} it is a helper function to multiply matrices
with vector by rows or columns
}
\note{
This function multiplies matrices with vectors by rows or colums.
If by = "row" all values of each row will be multiplied with each value
of the vector x.
If by = "col" all values of each column will be multiplied with each value
of the vector x.
}
\examples{
\dontrun{
# Do not run
data(net)
veh <- age_ldv(net$ldv[1:4], agemax = 4)
matvect(veh, 1:4)
}
}
