% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_ef.R
\name{moves_ef}
\alias{moves_ef}
\title{MOVES emission factors}
\usage{
moves_ef(
  ef,
  vehicles,
  source_type_id = 21,
  process_id = 1,
  fuel_type_id = 1,
  pollutant_id = 2,
  road_type_id = 5,
  speed_bin
)
}
\arguments{
\item{ef}{emission factors from EmissionRates_running exported from MOVES}

\item{vehicles}{Name of category, with length equal to fuel_type_id and other with id}

\item{source_type_id}{Number to identify type of vehicle as defined by MOVES.}

\item{process_id}{Number to identify emission process defined by MOVES.}

\item{fuel_type_id}{Number to identify type of fuel as defined by MOVES.}

\item{pollutant_id}{Number to identify type of pollutant as defined by MOVES.}

\item{road_type_id}{Number to identify type of road as defined by MOVES.}

\item{speed_bin}{Data.frame or vector of avgSpeedBinID as defined by MOVES.}
}
\value{
EmissionFactors data.frame
}
\description{
\code{\link{moves_ef}} reads and filter MOVES
data.frame of emission factors.
}
\note{
`decoder` shows a decoder for MOVES to identify
}
\examples{
{
data(decoder)
decoder
}
}
