% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_exhaust.R
\name{moves_exhaust}
\alias{moves_exhaust}
\title{Estimation of exhaust emissions using MOVES}
\usage{
moves_exhaust(
  veh,
  lkm,
  ef,
  sourceTypeID = 21,
  fuelTypeID = 1,
  pollutantID = 91,
  roadTypeID = 5,
  fuel_type,
  speed_bin,
  profile,
  vehicle = NULL,
  vehicle_type = NULL,
  fuel_subtype = NULL,
  path_all,
  verbose = FALSE
)
}
\arguments{
\item{veh}{"Vehicles" data-frame or list of "Vehicles" data-frame. Each data-frame
as number of columns matching the age distribution of that ype of vehicle.
The number of rows is equal to the number of streets link.}

\item{lkm}{Length of each link in miles}

\item{ef}{emission factors from EmissionRates_running exported from MOVES}

\item{sourceTypeID}{Number to identify type of vehicle as defined by MOVES.}

\item{fuelTypeID}{Number to identify type of fuel as defined by MOVES.}

\item{pollutantID}{Number to identify type of pollutant as defined by MOVES.}

\item{roadTypeID}{Number to identify type of road as defined by MOVES.}

\item{fuel_type}{Data.frame of fuelSubtypeID exported by MOVES.}

\item{speed_bin}{Data.frame or vector of avgSpeedBinID as defined by MOVES.}

\item{profile}{Data.frame or Matrix with nrows equal to 24 and ncol 7 day of
the week}

\item{vehicle}{Character, type of vehicle}

\item{vehicle_type}{Character, subtype of vehicle}

\item{fuel_subtype}{Character, subtype of vehicle}

\item{path_all}{Character to export whole estimation. It is not recommended since it
is usually too heavy.}

\item{verbose}{Logical; To show more information. Not implemented yet}
}
\value{
a list with emissions at each street and data.base aggregated by categories. See \code{link{emis_post}}
}
\description{
\code{\link{moves_exhaust}} estimates vehicular emissions as the product of the
vehicles on a road, length of the road, emission factor avaliated at the
respective speed. \eqn{E = VEH*LENGTH*EF(speed)}
}
\note{
`decoder` shows a decoder for MOVES to identify
}
\examples{
\dontrun{
data(decoder)
decoder
}
}
