% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\docType{data}
\name{net}
\alias{net}
\title{Road network of the west part of Sao Paulo city}
\format{A data frame with 1796 rows and 1 variables:
\describe{
  \item{highway}{Type of streets}
}}
\source{
\url{http://www.cetsp.com.br/}
}
\usage{
data(net)
}
\description{
This dataset is a SpatialLineDataFrame of sp package with roads
from a traffic simulations made by CET Sao Paulo, Brazil
}
\keyword{datasets}
