% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_cold.R
\docType{data}
\name{pc_cold}
\alias{pc_cold}
\title{Profile of Vehicle start patterns}
\format{
A data frame with 24 rows and 1 variables:
\describe{
  \item{V1}{24 hours profile vehicle starts for Monday}
}
}
\usage{
data(pc_cold)
}
\description{
This dataset is a dataframe with percetage of hourly starts with a lapse
of 6 hours with engine turned off. Data source is:
Lents J., Davis N., Nikkila N., Osses M. 2004. Sao Paulo vehicle activity
study. ISSRC. www.issrc.org
}
\keyword{datasets}
