% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_profile.R
\docType{data}
\name{pc_profile}
\alias{pc_profile}
\title{Profile of traffic data 24 hours 7 n days of the week}
\format{
A data frame with 24 rows and 7 variables:
\describe{
  \item{V1}{24 hours profile for Monday}
  \item{V2}{24 hours profile for Tuesday}
  \item{V3}{24 hours profile for Wednesday}
  \item{V4}{24 hours profile for Thursday}
  \item{V5}{24 hours profile for Friday}
  \item{V6}{24 hours profile for Saturday}
  \item{V7}{24 hours profile for Sunday}
}
}
\usage{
data(pc_profile)
}
\description{
This dataset is a dataframe with traffic activity normalized
monday 08:00-09:00. This data is normalized at 08:00-09:00.
It comes from data of toll stations near Sao Paulo City.
The source is ARTESP (www.artesp.com.br)
}
\keyword{datasets}
