% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.EmissionFactors.R
\name{plot.EmissionFactors}
\alias{plot.EmissionFactors}
\title{Plot of EmissionFactors}
\usage{
plot.EmissionFactors(ef, xlab = "", ...)
}
\arguments{
\item{ef}{Object with class "EmissionFactors"}

\item{xlab}{Labs for horizontal axe}
}
\value{
Plot Speed class
}
\description{
The object with class "EmissionFactors" is esentially a
data.frame with columns as type of vehicles and rows as age. Therefore plot
seeks to inform the emission factor producing a panel of plots of emission
factors, each panel to each type of vehicle. When the emission factors
are functions, other methods should be used.
}
\examples{
\dontrun{
data(fe2015)
names(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- as.EmissionFactors(df)
plot(ef1)
}
}
