% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Vehicles.R
\name{plot.Vehicles}
\alias{plot.Vehicles}
\title{plot}
\usage{
plot.Vehicles(veh, by = "col", mean = TRUE, xlab = "Index", ...)
}
\arguments{
\item{veh}{An object with class "Vehicles"}

\item{by}{Character that determines the type of plot. It accept the values:
"col", "streets" and "default". When by is "col" it is shown a plot of the
sum of the columns (See \code{\link{colSums}}). When by is "streets"
performs  sum of the streets via \code{\link{rowSums}}.When by is "default"
it plots the default method for plot (See \code{\link{plot}}).}

\item{mean}{a logical value. When mean is TRUE and by is "col" it adds
the weighted mean with and when by is "streets" the mean.}

\item{xlab}{xlab}

\item{...}{ignored}
}
\description{
The object with class "Vehicles" is esentially a data.frame
with columns as type of vehicles and rows as streets. Therefore the plot
seeks to inform the total number by type of vehicles
}
\details{
Plot of Vehicles
}
\examples{
\dontrun{
# Do not run
data(net)
LT_B5 <- age_hdv(x = net$ldv,name = "LT_B5")
graphics::plot(colSums(LT_B5), xlab = "c")
}
}
