% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EmissionFactorsList.R
\name{print.EmissionFactorsList}
\alias{print.EmissionFactorsList}
\title{Printing class "EmissionFactorsList"}
\usage{
print.EmissionFactorsList(ef, default = F, ...)
}
\arguments{
\item{ef}{Object with class "EmissionFactors"}

\item{default}{when T call method print.listof \code{\link{print.listof}}.
When F applies returns messages with type of nested lists. In vein it is
assumed that "EmissionFactorsList" are list of functions. The number
of elements in the list is the number of function that apply to a specific
type of vehicle. For example, an "EmissionFactorsList" with 30 functions
applies to a "Vehicles" data.frame with 30 categories. The function
\code{\link{emis}} relates both type of objects.}
}
\value{
Print method
}
\description{
A simple method for printing objects with class
"EmissionFactorsList".
}
\examples{
\dontrun{
# Do not run
data(fe2015)
names(fe2015)
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef2 <- as.EmissionFactors(df, lfx = T)
ef2
ef2[[1]]
}
}
\seealso{
\code{\link{emis}}
}
