% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EmissionsArray.R
\name{EmissionsArray}
\alias{EmissionsArray}
\alias{print.EmissionsArray}
\title{Printing class "EmissionsArray"}
\usage{
EmissionsArray(e, ...)

\method{print}{EmissionsArray}(e, default = F)
}
\arguments{
\item{e}{Object with class "EmissionsArray"}

\item{...}{ignored}

\item{default}{when T call method print.default
\code{\link{print.default}}. When F returns messages with array structure}
}
\value{
Print method
}
\description{
A simple method for printing objects with class "EmissionsArray"
}
\examples{
\dontrun{
# Do not run
}
}
