% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EmissionsList.R
\name{print.EmissionsList}
\alias{print.EmissionsList}
\title{Printing class "Emissionslist"}
\usage{
print.EmissionsList(e, default = F, ...)
}
\arguments{
\item{e}{Object with class "EmissionsList"}

\item{default}{when T call method print.listof
\code{\link{print.listof}}. When F returns messages with list structure}
}
\value{
Print method
}
\description{
A simple method for printing objects with class "EmissionsList".
It prints \code{\link{head}} of the first numeric element of the list.
}
\examples{
\dontrun{
# Do not run
}
}
