% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Speed.R
\name{print.Speed}
\alias{print.Speed}
\title{Printing class "Speed"}
\usage{
print.Speed(spd, all = TRUE, ...)
}
\arguments{
\item{spd}{Object with class "Speed"}

\item{all}{when T call method print.data.frame \code{\link{print.data.frame}}.
When F applies a  default print to each column}
}
\value{
Print method
}
\description{
A simple method for printing objects with class "Speed".
}
\examples{
\dontrun{
# Do not run
data(net)
data(pc_profile)
pcw <- temp_fact(net$ldv+net$hdv, pc_profile)
class(pcw)
df <- netspeed(pcw, net$ps,net$ffs, net$capacity, net$lkm, alpha = 1,isList=F)
class(df)
df
}
}
