% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Vehicles.R
\name{print.Vehicles}
\alias{print.Vehicles}
\title{Printing class "Vehicles"}
\usage{
print.Vehicles(veh, all = TRUE, ...)
}
\arguments{
\item{veh}{Objecto with class "Vehicles"}

\item{all}{when T call method print.data.frame \code{\link{print.data.frame}}.}
}
\value{
Print method
When F applies a  default print to each column
}
\description{
A simple method for printing objects with class "Vehicles".
}
\examples{
\dontrun{
# Do not run
data(net)
LT_B5 <- age_hdv(x = lt,name = "LT_B5")
LT_B5
}
}
