% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_soak.R, R/running_losses.R
\name{running_losses-deprecated}
\alias{running_losses-deprecated}
\title{Estimation of average daily  hot-soak evaporative emissions}
\arguments{
\item{eshotc}{average daily hot-soak evaporative factor for vehicles with
carburetor or fuel return system}

\item{eswarmc}{average daily cold-warm-soak evaporative factor for vehicles
with carburator or fuel return system}

\item{eshotfi}{average daily hot-soak evaporative factor for vehicles
with fuel injection and returnless fuel systems}

\item{x}{Numeric or data.frame; Mean number of trips per vehicle per day or Vehicles.}

\item{erhotfi}{average daily hot running losses evaporative factor
for vehicles with fuel injection and returnless fuel systems}

\item{lkm}{Numeric with class units and unit km.}

\item{carb}{fraction of gasoline vehicles with carburetor or fuel return system.}

\item{p}{Fraction of trips finished with hot engine}

\item{erhotc}{average daily running losses evaporative factor for vehicles with
carburator or fuel return system}

\item{erwarmc}{average daily cold and warm running losses evaporative factor
for vehicles with carburetor or fuel return system}
}
\value{
numeric vector of emission estimation in grams

numeric vector of emission estimation in grams
}
\description{
\code{hot_soak} estimates of evaporative emissions from EMEP/EEA
emisison guidelines

\code{running_losses} estimates evaporative emissions from
EMEP/EEA emisison guidelines
}
\seealso{
\code{\link{vein-deprecated}}

\code{\link{vein-deprecated}}
}
\keyword{internal}
