% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speed.R
\name{Speed}
\alias{Speed}
\title{Speed}
\usage{
Speed(spd, ...)
}
\arguments{
\item{spd}{Object which can be "matrix", "data.frame" or "numeric"}

\item{...}{ignored}
}
\value{
Objects of class "Speed"
}
\description{
Generic function for class "Speed"
}
\details{
Generic function for class "Speed"
}
