% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_moves.R
\name{speed_moves}
\alias{speed_moves}
\title{Return speed bins according to US/EPA MOVES model}
\usage{
speed_moves(x, net)
}
\arguments{
\item{x}{Object with class, "sf",  "data.frame", "matrix" or "numeric" with speeds
in miles/h (mph)}

\item{net}{optional  spatial dataframe of class "sf".
it is transformed to "sf".#' @importFrom sf st_sf st_set_geometry st_geometry}
}
\description{
\code{speed_moves} return an object of average speed bins as defined by
US EPA MOVES. The input must be speed as miles/h (mph)
}
\examples{
{
data(net)
net$mph <- units::set_units(net$ps, "miles/h")
net$speed_bins <- speed_moves(net$mph)
head(net)
speed_moves(net["ps"])
}
}
