% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.EmissionsList.R
\name{summary.EmissionsList}
\alias{summary.EmissionsList}
\title{Summary of EmissionsList}
\usage{
summary.EmissionsList(e, default = F)
}
\arguments{
\item{e}{Objecto of class "EmissionsArray"}

\item{default}{Logical value. Whe  default is TRUE it returns the default
summary. When FALSE it return the structure of the list}
}
\value{
Summary for class "EmissionsList"
}
\description{
The object with class "EmissionsList" is list of emissions.
The summary method consists in plotting a mesage with the structure of
the list and indicating the summary of the first numeric element
}
\examples{
\dontrun{
#Do not run
}
}
\seealso{
\code{\link{head}}
}
