% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Speed.R
\name{Speed}
\alias{Speed}
\alias{summary.Speed}
\title{Summary of Speed}
\usage{
Speed(spd, ...)

summary.Speed(spd, by = "col")
}
\arguments{
\item{spd}{Object of class Speed}

\item{...}{ignored}

\item{by}{This parameter allows 4 values: "col", "streets", "all" and
"default" "col" returns a summary of the sum of the columns
via \code{\link{colSums}}, "streets" summary of the sum of the streets
via \code{\link{rowSums}}, "all" returns a simple summary to each combunation
columns x street and "default" the default summary.data.frame via
\code{\link{summary.data.frame}}}
}
\value{
Summary for vehicle classes
}
\description{
The object with class "Speed" is esentially a data.frame
with columns as type of vehicles and rows as streets. The summary method
considers this by arguments by, mean and default to produce summaries
the information of vehicles in a helpful way.
}
\examples{
\dontrun{
data(net)
data(pc_profile)
pcw <- temp_fact(net$ldv+net$hdv, pc_profile)
class(pcw)
df <- netspeed(pcw, net$ps,net$ffs, net$capacity, net$lkm, alpha = 1,isList=F)
class(df)
summary(df, by="age")
summary(df, by="streets")
summary(df, by="all")
summary(df, by="default")
}
}
\seealso{
\code{\link{summary}}
}
