% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Vehicles.R
\name{summary.Vehicles}
\alias{summary.Vehicles}
\title{Summary of vein classes}
\usage{
summary.Vehicles(object, ...)
}
\arguments{
\item{object}{An object with class  of vein}

\item{...}{ignored}
}
\value{
Summary for vein classes
}
\description{
The summary methods
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
\code{\link{summary}}
}
