% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vehicles.R
\name{Vehicles}
\alias{Vehicles}
\title{Vehicles}
\usage{
Vehicles(veh, ...)
}
\arguments{
\item{veh}{Object which can be "matrix", "data.frame" or "numeric"}

\item{...}{ignored}
}
\value{
Objects of class "Vehicles"
}
\description{
Generic function for class "Vehicles"
}
\details{
Generic function for class "Vehicles"
}
