% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein-package.R
\docType{package}
\name{vein-package}
\alias{vein}
\alias{vein-package}
\title{vein: Vehicular Emissions Inventories}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Elaboration of vehicular emissions inventories, consisting in four stages, pre-processing activity data, preparing emissions factors, estimating the emissions and post-processing of emissions in maps and databases. More details in Ibarra-Espinosa et al (2018) \doi{10.5194/gmd-11-2209-2018}. Before using VEIN you need to know the vehicular composition of your study area, in other words, the combination of type of vehicles, size and fuel of the fleet. Then, it is recommended to start with the project to download a template to create a structure of directories and scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoschem/vein}
  \item Report bugs at \url{https://github.com/atmoschem/vein/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{zergioibarra@gmail.com} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Authors:
\itemize{
  \item Sergio Ibarra-Espinosa \email{zergioibarra@gmail.com} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Schuch \email{underschuch@gmail.com} (\href{https://orcid.org/0000-0001-5977-4519}{ORCID}) [contributor]
  \item Joao Bazzo \email{joao.bazzo@gmail.com} (\href{https://orcid.org/0000-0002-7371-1116}{ORCID}) [contributor]
  \item Mario Gavidia-Calderón \email{mario.calderon@iag.usp.br} (\href{https://orcid.org/0000-0003-4536-5006}{ORCID}) [contributor]
  \item Karl Ropkins \email{k.ropkins@its.leeds.ac.uk} (\href{https://orcid.org/0000-0002-0294-6997}{ORCID}) [contributor]
}

}
\keyword{internal}
