% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly.R
\name{weekly}
\alias{weekly}
\title{Reordene the emission for simulation}
\format{Emissions}
\usage{
weekly(EMISSION, start = "mon", hours = 168, verbose = FALSE)
}
\arguments{
\item{EMISSION}{GriddedEmissionsArray output}

\item{start}{Date or the start weekday first 3 letters}

\item{hours}{number of hours needed to the simualation}

\item{verbose}{display additional information}
}
\description{
returns the emission array matching with corresponding weekdays and with
the desired number of hours, recycling or droping hours from the emission array.
}
\examples{
\dontrun{
wCO <- weekly(CO, start = "sat", hours = 24, verbose = TRUE)
wCO <- weekly(CO, start = as.Date("2016-04-06"), hours = 241, verbose = TRUE)
}

}
\author{
Daniel Schuch
}
