% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lkm.R
\name{add_miles}
\alias{add_miles}
\title{Construction function to add unit miles}
\usage{
add_miles(x)
}
\arguments{
\item{x}{Object with class "data.frame", "matrix",  "numeric" or "integer"}
}
\value{
Objects of class "data.frame" or "units"
}
\description{
\code{add_miles} just  add unit 'miles' to different R objects
}
\examples{
\dontrun{
a <- add_miles(rnorm(100)*10)
plot(a)
b <- add_miles(matrix(rnorm(100)*10, ncol = 10))
print(head(b))
}
}
\seealso{
Other Add distance unitts: 
\code{\link{add_lkm}()}
}
\concept{Add distance unitts}
