% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_hdv.R
\name{age_hdv}
\alias{age_hdv}
\title{Returns amount of vehicles at each age}
\usage{
age_hdv(
  x,
  name = "age",
  a = 0.2,
  b = 17,
  agemin = 1,
  agemax = 50,
  k = 1,
  bystreet = F,
  net,
  verbose = FALSE,
  namerows,
  time
)
}
\arguments{
\item{x}{Numeric; numerical vector of vehicles with length equal to lines features of road network}

\item{name}{Character; of vehicle assigned to columns of dataframe}

\item{a}{Numeric; parameter of survival equation}

\item{b}{Numeric; parameter of survival equation}

\item{agemin}{Integer; age of newest vehicles for that category}

\item{agemax}{Integer; age of oldest vehicles for that category}

\item{k}{Numeric; multiplication factor. If its length is > 1, it must match the length of x}

\item{bystreet}{Logical; when TRUE it is expecting that 'a' and 'b' are numeric vectors with length equal to x}

\item{net}{SpatialLinesDataFrame or Spatial Feature of "LINESTRING"}

\item{verbose}{Logical;  message with average age and total numer of vehicles}

\item{namerows}{Any vector to be change row.names. For instance, name of
regions or streets.}

\item{time}{Character to be the time units as denominator, eg "1/h"}
}
\value{
dataframe of age distrubution of vehicles at each street
}
\description{
\code{\link{age_hdv}} returns amount of vehicles at each age
}
\note{
The functions age* produce distribution of the circulating fleet by age of use.
The order of using these functions is:

1. If you know the distribution of the vehicles by age of use , use:  \code{\link{my_age}}
2. If you know the sales of vehicles, or the registry of new vehicles,
use \code{\link{age}} to apply a survival function.
3. If you know the theoretical shape of the circulating fleet and you can use
\code{\link{age_ldv}}, \code{\link{age_hdv}} or \code{\link{age_moto}}. For instance,
you dont know the sales or registry of vehicles, but somehow you know
the shape of this curve.
4. You can use/merge/transform/adapt any of these functions.
}
\examples{
{
data(net)
LT_B5 <- age_hdv(x = net$hdv,name = "LT_B5")
plot(LT_B5)
LT_B5 <- age_hdv(x = net$hdv, name = "LT_B5", net = net)
plot(LT_B5)
}
}
