% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nt.R
\name{check_nt}
\alias{check_nt}
\title{Check the number of threads}
\usage{
check_nt()
}
\value{
Integer with the maxnumber of threads
}
\description{
\code{check_nt} check the number of threads in this machine
}
\examples{
{
check_nt()
}
}
