% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cold_mileage.R
\name{cold_mileage}
\alias{cold_mileage}
\title{Fraction of mileage driven with a cold engine or catalizer below normal temperature}
\usage{
cold_mileage(ltrip, ta)
}
\arguments{
\item{ltrip}{Numeric; Length of trip.}

\item{ta}{Numeric; average monthly temperature}
}
\description{
This function depends length of trip and on ambient temperature.
From the guidelines  EMEP/EEA air pollutant emission inventory guidebook
http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook
}
\examples{
{
cold_mileage(10, 0)
}
}
\keyword{cold}
\keyword{mileage}
