% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmonth.R
\name{dmonth}
\alias{dmonth}
\title{Number of days of the month}
\usage{
dmonth(year, month)
}
\arguments{
\item{year}{Numeric}

\item{month}{Numeric}
}
\value{
days of the month
}
\description{
\code{\link{ef_ldv_speed}} return the number of days of the month
}
\examples{
\dontrun{
dmonth(2022, 1)
dmonth(Sys.Date())
}
}
\seealso{
Other helpers: 
\code{\link{colplot}()},
\code{\link{to_latex}()}
}
\concept{helpers}
