% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_china.R
\name{ef_china_h}
\alias{ef_china_h}
\title{Correction of Chinese factors by altitude}
\usage{
ef_china_h(h, v = "PV", t = "Small", f = "G", standard, p)
}
\arguments{
\item{h}{numeric aktitude}

\item{v}{Character; category vehicle: "PV" for Passenger Vehicles or 'Trucks"}

\item{t}{Character; sub-category of of vehicle: PV Gasoline:  "Mini", "Small","Medium",
"Large", "Taxi", "Motorcycles", "Moped", PV Diesel: "Mediumbus", "Largebus",
 "3-Wheel". Trucks: "Mini", "Light" , "Medium", "Heavy"}

\item{f}{Character;fuel: "G", "D", "CNG"}

\item{standard}{Character vector; "PRE", "I", "II", "III", "IV", "V".}

\item{p}{Character; pollutant: "CO", "NOx","HC", "PM", "Evaporative_driving"
or "Evaporative_parking"}
}
\value{
long data.frame
}
\description{
Correction of Chinese emission
}
\examples{
{
ef_china_h(h = 1600, standard = "I", p = "CO")
}
}
\seealso{
Other China: 
\code{\link{ef_china_det}()},
\code{\link{ef_china_hu}()},
\code{\link{ef_china_long}()},
\code{\link{ef_china_speed}()},
\code{\link{ef_china_te}()},
\code{\link{ef_china_th}()},
\code{\link{ef_china}()}
}
\concept{China}
