% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_emfac.R
\name{ef_emfac}
\alias{ef_emfac}
\title{Emission Factors from EMFAC emission factors}
\usage{
ef_emfac(efpath)
}
\arguments{
\item{efpath}{Character path to EMFAC ef (g/miles)}
}
\value{
data.table with emission estimation in long format
}
\description{
\code{\link{ef_emfac}} reads path to ef EMFAC.
You must download the
emission factors from EMFAC website.
}
\examples{
\dontrun{
# do not run
}
}
