% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_ldv_cold.R
\name{ef_ldv_cold}
\alias{ef_ldv_cold}
\title{Cold-Start Emissions factors for Light Duty Vehicles}
\usage{
ef_ldv_cold(v = "LDV", ta, cc, f, eu, p, k = 1,
  show.equation = FALSE)
}
\arguments{
\item{v}{Character; Category vehicle:  "LDV"}

\item{ta}{Numeric; Ambient temperature. Monthly mean can be used}

\item{cc}{Character; Size of engine in cc: "<=1400",  "1400_2000" or ">2000"}

\item{f}{Character; Type of fuel: "G", "D" or "LPG"}

\item{eu}{Character; Euro standard: "PRE", "I", "II", "III",  "IV", "V", "VI" or "VIc"}

\item{p}{Character; Pollutant: "CO", "FC", "NOx", "HC" or "PM"}

\item{k}{Numeric; Multiplication factor}

\item{show.equation}{Option to see or not the equation parameters}
}
\value{
an emission factor function which depends of the average speed V
and ambient temperature. g/km
}
\description{
This function returns speed functions which depends on ambient temperature
average speed. The emission factors comes from the guidelines  EMEP/EEA air pollutant
emission inventory guidebook
http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook
}
\examples{
{
ef1 <- ef_ldv_cold(ta = 15, cc = "<=1400", f ="G", eu = "PRE", p = "CO")
ef1(10)
# lets create a matrix of ef cold at different speeds and temperatures
te <- -50:50
lf <- sapply(1:length(te), function(i){
ef_ldv_cold(ta = te[i], cc = "<=1400", f ="G", eu = "I", p = "CO")(1:120)
})
for(i in 1:ncol(lf)){
lf[, i] <- ifelse(lf[, i] < 0, 0, lf[, i])
}
filled.contour(lf, col= cptcity::lucky())
}
}
\keyword{cold}
\keyword{emission}
\keyword{factors}
