% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_chem.R
\name{emis_chem}
\alias{emis_chem}
\title{Aggregate emissions by lumped groups in chemical mechanism}
\usage{
emis_chem(dfe, mechanism, colby)
}
\arguments{
\item{dfe}{data.frame with column `emissions` in grams and `pollutant`.}

\item{mechanism}{Character, any of "SAPRC", "RACM", "RADM2", "CBMZ",
"MOZART", "SAPRC99", "CB05", "CB06CMAQ", "RACM2CMAQ", "SAPRC99CMAQ",
"SAPRC07CMAQ",  "SAPRC07A","CRIMECH"
WRF: "RADM2_SORG", "CBMZ_MOSAIC", "CPTEC", "GOCART_CPTEC", "MOZEM",
"MOZCEM", "CAMMAM", "MOZMEM", "MOZC_T1_EM", "CB05_OPT1" or "CB05_OPT2"}

\item{colby}{Character indicating column name for aggregating extra column.
For instance, region or province}
}
\value{
data.frame with lumped groups by chemical mechanism. It transform
emissions in grams to mol.
}
\description{
\code{\link{emis_chem}} aggregates emissions by chemical mechanism
and convert grams to mol. This function reads all hydrocarbos and respective
criteria polluants specified in \code{\link{ef_ldv_speed}} and \code{\link{ef_hdv_speed}}.
}
\note{
This feature is experimental and the mapping of pollutants and lumped
species may change in future.
This function is converting the intial data.frame input into data.table
}
\examples{
{
# CO
df <- data.frame(emission = Emissions(1:10))
df$pollutant = "CO"
emis_chem(df, "CBMZ")
# hexanal
df$pollutant = "hexanal"
emis_chem(df, "CBMZ")
# propadiene and NO2
df2 <- df1 <- df
df1$pollutant = "propadiene"
df2$pollutant = "NO2"
(dfe <- rbind(df1, df2))
emis_chem(dfe, "CBMZ")
dfe$region <- rep(letters[1:2], 10)
emis_chem(dfe, "CBMZ", "region")
}
}
\seealso{
\code{\link{ef_ldv_speed}} \code{\link{ef_hdv_speed}} \code{\link{speciate}}
}
