% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_china.R, R/emis_long.R
\name{emis_china}
\alias{emis_china}
\alias{emis_long}
\title{Estimation with Chinese factors}
\usage{
emis_china(
  x,
  lkm,
  tfs,
  v = "PV",
  t = "Small",
  f = "G",
  standard,
  s,
  speed,
  te,
  hu,
  h,
  yeardet = 2016,
  p,
  verbose = TRUE,
  array = FALSE
)

emis_long(x, lkm, ef, tfs, speed, verbose = TRUE, array = FALSE)
}
\arguments{
\item{x}{Vehicles data.frame. x repeats down for each hour}

\item{lkm}{Length of each link in km. lkm repeats down for each hour}

\item{tfs}{temporal factor}

\item{v}{Character; category vehicle: "PV" for Passenger Vehicles or 'Trucks"}

\item{t}{Character; sub-category of of vehicle: PV Gasoline:  "Mini", "Small","Medium",
"Large", "Taxi", "Motorcycles", "Moped", PV Diesel: "Mediumbus", "Largebus",
 "3-Wheel". Trucks: "Mini", "Light" , "Medium", "Heavy"}

\item{f}{Character;fuel: "G", "D", "CNG", "ALL"}

\item{standard}{Character vector; "PRE", "I", "II", "III", "IV", "V".}

\item{s}{Sulhur in ppm}

\item{speed}{Speed data.frame (nrow x)}

\item{te}{Temperature (length tfs)}

\item{hu}{Humidity (length tfs)}

\item{h}{Altitude (length nrow x)}

\item{yeardet}{Year, default 2016}

\item{p}{Character; pollutant: "CO", "NOx","HC", "PM", "Evaporative_driving"
or "Evaporative_parking"}

\item{verbose}{Logical to show more info}

\item{array}{Logical to return EmissionsArray or not}

\item{ef}{data.frame. ef repeats down for each hour}
}
\value{
long data.frame

long data.frame
}
\description{
Emissions estimates

Emissions estimates
}
\examples{
{
ef_china_h(h = 1600, p = "CO")
}
{
data(net)
net <- net[1:100, ]
data(pc_profile)
x <- age_ldv(net$ldv)
pc_week <- temp_fact(net$ldv+net$hdv, pc_profile[[1]])
df <- netspeed(pc_week,
               net$ps,
               net$ffs,
               net$capacity,
               net$lkm,
               alpha = 1)

s  <- do.call("rbind",lapply(1:ncol(df), function(i) {
 as.data.frame(replicate(ncol(x), df[, i]))
}))

ef <- ef_wear(wear = "tyre",
              type = "PC",
              pol = "PM10",
              speed = as.data.frame(s))

e <- emis_long(x = x,
               lkm = net$lkm,
               ef = ef,
               tfs = pc_profile[[1]],
               speed = df)

ae <- emis_long(x = x,
               lkm = net$lkm,
               ef = ef,
               tfs = pc_profile[[1]],
               speed = df,
               array = TRUE)
}
}
\seealso{
Other China: 
\code{\link{ef_china_det}()},
\code{\link{ef_china_hu}()},
\code{\link{ef_china_h}()},
\code{\link{ef_china_long}()},
\code{\link{ef_china_speed}()},
\code{\link{ef_china_s}()},
\code{\link{ef_china_te}()},
\code{\link{ef_china_th}()},
\code{\link{ef_china}()}

Other China: 
\code{\link{ef_china_det}()},
\code{\link{ef_china_hu}()},
\code{\link{ef_china_h}()},
\code{\link{ef_china_long}()},
\code{\link{ef_china_speed}()},
\code{\link{ef_china_s}()},
\code{\link{ef_china_te}()},
\code{\link{ef_china_th}()},
\code{\link{ef_china}()}
}
\concept{China}
