% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_grid.R
\name{emis_grid}
\alias{emis_grid}
\title{Allocate emissions into a grid}
\usage{
emis_grid(spobj, g, sr, type = "lines")
}
\arguments{
\item{spobj}{A spatial dataframe of class sp}

\item{g}{Grid}

\item{sr}{Spatial reference, default is "+init=epsg:4326"}

\item{type}{type of geometry: "lines" or "points"}
}
\description{
The allocation is proportionally to each grid cell. The process is
performed by intersection between geometries and the grid. Geometries
suported, so, far are lines with raster::intersect and points with
sp::over. The allocation of lines is by interaction, then update the
pollutant values according the new length of road inside each grid cell.
It means that requires "sr" according with your location for the projection.
It is assumed that soobj is a spatial*DataFrame with the pollutant in data.
Also, it is required that, when is a SpatialLinesDataFrame, there is a field
called lkm, with the length of the road, in this case, in km.
}
\examples{
\dontrun{
#do not run
data(net)
net@data$POL <- rnorm(n = nrow(net@data), mean = 1000, sd = 100)
grid <- make_grid(net, width = 0.5/102.47, height = 0.5/102.47) #500 mts
gPOL <- emis_grid(spobj = net, g = grid, pol = "POL", sr = "+init=epsg:29183")
spplot(gPOL, "POL", scales=list(draw=T),
       sp.layout = list("sp.lines", net, pch = 16, cex = 2, col = "black"))
}
}
