% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactors.R
\name{EmissionFactors}
\alias{EmissionFactors}
\title{Generic function EmissionFactors}
\usage{
EmissionFactors(ef, ...)
}
\arguments{
\item{ef}{object}

\item{...}{ignored}
}
\description{
Generic function
}
