% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactors.R, R/EmissionFactorsList.R
\name{EmissionFactors}
\alias{EmissionFactors}
\alias{EmissionFactors}
\title{Generic function of class "EmissionFactors"}
\usage{
EmissionFactors(ef, ...)

EmissionFactors(ef, ...)
}
\value{
Construtor of objects with class "EmissionFactors"

Construtor of objects with class "EmissionFactorsList"
}
\description{
Returns object with class "EmissionFactors"

Returns object with class "EmissionFactorsList"
}
