% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactors.R
\name{EmissionFactors}
\alias{EmissionFactors}
\alias{print.EmissionFactors}
\alias{summary.EmissionFactors}
\alias{plot.EmissionFactors}
\title{Construction function for class "EmissionFactors"}
\usage{
EmissionFactors(x, ...)

\method{print}{EmissionFactors}(x, ...)

\method{summary}{EmissionFactors}(object, ...)

\method{plot}{EmissionFactors}(x, ...)
}
\arguments{
\item{x}{Object with class "data.frame", "matrix" or "numeric"}

\item{...}{ignored}

\item{object}{Object with class "EmissionFactors"}
}
\value{
Objects of class "EmissionFactors" or "units"
}
\description{
\code{EmissionFactors} returns a tranformed object with class
"EmissionFactors" and units g/km.
}
\examples{
{
data(fe2015)
names(fe2015)
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- EmissionFactors(df)
class(ef1)
summary(ef1)
plot(ef1)
print(ef1)
}
}
