% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactorsList.R
\name{EmissionFactorsList}
\alias{EmissionFactorsList}
\title{EmissionFactors}
\usage{
EmissionFactorsList(ef, ...)
}
\arguments{
\item{ef}{Object which can be "list" with "numeric" elements}

\item{...}{ignored}
}
\value{
Objects of class "EmissionFactorsList" or "units"
}
\description{
Generic function for class "EmissionFactorsList"
}
\details{
Generic function for lass "EmissionFactorsList"
}
