% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionFactorsList.R
\name{EmissionFactorsList}
\alias{EmissionFactorsList}
\alias{print.EmissionFactorsList}
\alias{summary.EmissionFactorsList}
\alias{plot.EmissionFactorsList}
\alias{print.EmissionFactorsList}
\alias{summary.EmissionFactorsList}
\alias{plot.EmissionFactorsList}
\title{Construction function for class "EmissionFactorsList"}
\usage{
EmissionFactorsList(x, ...)

\method{print}{EmissionFactorsList}(x, ...)

\method{summary}{EmissionFactorsList}(object, ...)

\method{plot}{EmissionFactorsList}(x, ...)
}
\arguments{
\item{x}{Object with class "list"}

\item{...}{ignored}

\item{object}{Object with class "EmissionFactorsList"}
}
\value{
Objects of class "EmissionFactorsList"
}
\description{
Returns a tranformed object with class"EmissionsFactorsList".
This functions has arguments to change of the numeric elements of the list.
}
\examples{
\dontrun{
data(fe2015)
names(fe2015
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- as.EmissionFactors(df)
class(ef1)
head(ef1)
ef2 <- as.EmissionFactors(df, lfx = T)
class(ef2)
class(ef2[[1]])
}
}
