% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{Emissions}
\alias{Emissions}
\title{Generic function of class "Emissions"}
\usage{
Emissions(veh, ...)
}
\value{
Construtor of objects with class "Emissions"
}
\description{
Returns object with class "Emissions"
}
