% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{Emissions}
\alias{Emissions}
\title{Emissions}
\usage{
Emissions(e, ...)
}
\arguments{
\item{e}{Object which can be "matrix", "data.frame" or "numeric"}

\item{...}{ignored}
}
\value{
Objects of class "Emissions" or "units"
}
\description{
Generic function for class "Emissions"
}
\details{
Generic function for lass "Emissions"
}
