% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsArray.R
\name{EmissionsArray}
\alias{EmissionsArray}
\title{Generic function of class "EmissionsArray"}
\usage{
EmissionsArray(veh, ...)
}
\value{
Construtor of objects with class "EmissionsArray"
}
\description{
Returns object with class "EmissionsArray"
}
