% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsList.R
\name{EmissionsList}
\alias{EmissionsList}
\alias{print.EmissionsList}
\alias{summary.EmissionsList}
\alias{plot.EmissionsList}
\title{DEPRECATED Construction function for class "EmissionsList"
Not exported}
\usage{
EmissionsList(x, ...)

\method{print}{EmissionsList}(x, ...)

\method{summary}{EmissionsList}(object, ...)

\method{plot}{EmissionsList}(x, ...)
}
\arguments{
\item{x}{object with class "EmissionList"}

\item{...}{ignored}

\item{object}{object with class "EmissionList"}
}
\value{
Objects of class "EmissionsList" and numeric elements as "units"
}
\description{
\code{EmissionsList} returns a tranformed object with class "EmissionsList".
}
\examples{
\dontrun{
# do not run
# DEPRECATED
}
}
