% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsList.R
\name{EmissionsList}
\alias{EmissionsList}
\title{Generic function of class "EmissionsList"}
\usage{
EmissionsList(veh, ...)
}
\value{
Construtor of objects with class "EmissionsList"
}
\description{
Returns object with class "EmissionsList"
}
