% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaporative.R
\name{Evaporative}
\alias{Evaporative}
\alias{print.Evaporative}
\alias{summary.Evaporative}
\alias{plot.Evaporative}
\title{DEPRECATED - Construction function for class "Evaporative"}
\usage{
Evaporative(x, ...)

\method{print}{Evaporative}(x, ...)

\method{summary}{Evaporative}(object, ...)

\method{plot}{Evaporative}(x, ...)
}
\arguments{
\item{x}{Object with class "numeric"}

\item{...}{ignored}

\item{object}{Object with class "Evaporative"}
}
\value{
Objects of class "Evaporative" or "units"
}
\description{
\code{Evaporative} returns a tranformed object with class "Evaporative" and
units g. This class represents the daily emissions presented by
Mellios G and Ntziachristos (2016) Gasoline evaporation, Tier 2. Eventually
it will be incorporated the techniques of Tier 3.
}
\examples{
\dontrun{
# do not run
# DEPRECATED
}
}
