% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project}
\alias{get_project}
\title{Download vein project to a specificor new directory}
\usage{
get_project(directory, case = "brasil", approach = "bottom-up")
}
\arguments{
\item{directory}{Character; Path to an existing or a new directory to be created.
It needs absolute path.}

\item{case}{Character; Currently only supports "brasil" (or "brazil").}

\item{approach}{Character; Currently only supports "bottom-up".}
}
\description{
\code{\link{get_project}} downloads a project for runnign vein.
The projects are available on Github.com/atmoschem/vein/projects
}
\examples{
\dontrun{
#do not run
}

}
