% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Creates rectangular grid for emission allocation}
\usage{
make_grid(spobj, width, height = width, polygon, crs = 4326, ...)
}
\arguments{
\item{spobj}{A spatial object of class sp or sf or a Character. When it is
a character, it is assumed that it is a path to wrfinput file to create
a grid class 'sf' based on this file. This is done by running eixport::wrf_grid.}

\item{width}{Width of grid cell. It is recommended to use projected values.}

\item{height}{Height of grid cell.}

\item{polygon}{Deprecated! \code{\link{make_grid}} returns only sf grid of
polygons.}

\item{crs}{coordinate reference system in numeric format from
http://spatialreference.org/ to transform/project spatial data using sf::st_transform.
The default value is 4326}

\item{...}{ignored}
}
\value{
A grid of polygons class 'sf'
}
\description{
\code{make_grid} creates a sf grid of polygons. The spatial
reference is taken from the spatial object.
}
\examples{
{
data(net)
grid <- make_grid(net, width = 0.5/102.47) #500 mts
plot(grid, axes = TRUE) #class sf
wrf <- paste(system.file("extdata", package = "eixport"),
"/wrfinput_d02", sep="")
gwrf  <- make_grid(wrf)
plot(gwrf, axes = TRUE)

}
}
