% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_age.R
\name{my_age}
\alias{my_age}
\title{Returns amount of vehicles at each age}
\usage{
my_age(x, y, name, k = 1)
}
\arguments{
\item{x}{numerical vector of vehicles}

\item{y}{Age dustribution of vehicles}

\item{name}{of vehicle assigned to columns of dataframe}

\item{k}{multiplication factor}
}
\value{
dataframe of age distrubution of vehicles
}
\description{
Returns amount of vehicles at each age using a numeric vector
}
\examples{
\dontrun{
# Do not run
pc <- rnorm(100, 300, 10)
dpc <- rnorm(10, 100, 1)
PC_E25_1400 <- my_age(x = pc,y = dpc, name = "PC_E25_1400")
}
}
