% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_age.R
\name{my_age}
\alias{my_age}
\title{Returns amount of vehicles at each age}
\usage{
my_age(x, y, name, k = 1, message = TRUE)
}
\arguments{
\item{x}{numerical vector of vehicles.}

\item{y}{Age dustribution of vehicles.}

\item{name}{of vehicle assigned to columns of dataframe.}

\item{k}{multiplication factor.}

\item{message}{message with average age and total numer of vehicles.}
}
\value{
dataframe of age distrubution of vehicles.
}
\description{
\code{my_age} returns amount of vehicles at each age using a
numeric vector.
}
\examples{
{
# Do not run
pc <- rnorm(100, 300, 10)
dpc <- c(rnorm(10, 99, 1), NA, NA, NA)
PC_E25_1400 <- my_age(x = pc, y = dpc, name = "PC_E25_1400")
plot(PC_E25_1400)
}
}
