% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\docType{data}
\name{net}
\alias{net}
\title{Road network of the west part of Sao Paulo city}
\format{
A Spatial data.frame (sf) with 1796 rows and 1 variables:
\describe{
  \item{ldv}{Light Duty Vehicles (veh/h)}
  \item{hdv}{Heavy Duty Vehicles (veh/h)}
  \item{lkm}{Length of the link (km)}
  \item{ps}{Peak Speed (km/h)}
  \item{ffs}{Free Flow Speed (km/h)}
  \item{tstreet}{Type of street}
  \item{lanes}{Number of lanes per link}
  \item{capacity}{Capacity of vehicles in each link (1/h)}
  \item{tmin}{Time for travelling each link (min)}
  \item{geometry}{geometry}
}
}
\source{
\url{https://www.cetsp.com.br/}
}
\usage{
data(net)
}
\description{
This dataset is a sf class object with roads
from a traffic simulations made by CET Sao Paulo, Brazil
}
\keyword{datasets}
