% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netspeed.R
\name{netspeed}
\alias{netspeed}
\title{Calculate speeds of traffic network}
\usage{
netspeed(q, ps, ffs, cap, lkm, alpha = 0.15, beta = 4, isList = FALSE,
  dist = "km", time = "h")
}
\arguments{
\item{q}{Data-frame of traffic flow to each hour (veh/h)}

\item{ps}{Peak speed (km/h)}

\item{ffs}{Free flow speed (km/h)}

\item{cap}{Capacity of link (veh/h)}

\item{lkm}{Distance of link (km)}

\item{alpha}{Parameter of BPR curves}

\item{beta}{Parameter of BPR curves}

\item{isList}{Logical to specify type of return, list or data-frame}
}
\value{
dataframe or list of speeds with units
}
\description{
Creates a dataframe of speeds fir diferent hours and each link based on
morning rush traffic data
}
\examples{
\dontrun{
# Do not run
data(net)
data(pc_profile)
qq <- as.matrix(net$ldv+net$hdv) \%*\% matrix(unlist(pc_profile), nrow=1)
df <- netspeed(qq, net$ps, net$ffs, net$capacity, net$lkm)
}
}
