% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Speed.R
\name{plot.Speed}
\alias{plot.Speed}
\title{Plot of Speed}
\usage{
plot.Speed(spd, by = "col", mean = TRUE, dist = "km", time = "h",
  xlab = NULL, ...)
}
\arguments{
\item{spd}{Object of class "Speed"}

\item{by}{Character that determines the type of plot. It accept the values:
"col", "streets" and "default".
When by is "col" it is shown a plot of the mean of the columns
(See \code{\link{colMeans}}) and also the standard deviation.
When by is "streets" performs  sum of the streets via \code{\link{rowSums}}.
When by is "default" it plots the default method for plot
(See \code{\link{plot.default}}).}

\item{mean}{a logical value. When mean is TRUE and by is "col" it adds
the weighted mean with and when by is "streets" the mean.}
}
\value{
Plot Speed class
}
\description{
The object with class "Speed" is esentially a data.frame
with columns as type of vehicles and rows as streets. Therefore the plot
seeks to inform the total number by type of vehicles. This functions
plot the mean speed and the standard deviation.
}
\examples{
\dontrun{
data(net)
data(pc_profile)
pcw <- temp_fact(net$ldv+net$hdv, pc_profile)
class(pcw)
df <- netspeed(pcw, net$ps,net$ffs, net$capacity, net$lkm, alpha = 1,isList=F)
class(df)
plot(df$S1)
plot(df, by = "col", mean=F, xlab="test")
plot(df, by="streets", mean=F)
plot(df, by="streets", mean=T)
}
}
