% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Vehicles.R
\name{plot.Vehicles}
\alias{plot.Vehicles}
\title{Plot of Vehicles}
\usage{
plot.Vehicles(veh, by = "col", mean = TRUE, ...)
}
\arguments{
\item{veh}{An object with class "Vehicles"}

\item{by}{Character that determines the type of plot. It accept the values:
"col", "streets" and "default".
When by is "col" it is shown a plot of the sum of the columns
(See \code{\link{colSums}}).
When by is "streets" performs  sum of the streets via \code{\link{rowSums}}.
When by is "default" it plots the default method for plot
(See \code{\link{plot.default}}).}

\item{mean}{a logical value. When mean is TRUE and by is "col" it adds
the weighted mean with and when by is "streets" the mean.}
}
\value{
Plot vehicles class
}
\description{
The object with class "Vehicles" is esentially a data.frame
with columns as type of vehicles and rows as streets. Therefore the plot
seeks to inform the total number by type of vehicles
}
\examples{
\dontrun{
# Do not run
data(net)
lt <- as.Vehicles(net$hdv)
class(lt)
plot(lt)
LT_B5 <- age_hdv(x = lt,name = "LT_B5")
plot(1:50,LT_B5[30,])
plot(LT_B5)
plot(LT_B5, by = "col", mean = T)
plot(LT_B5, by="streets")
plot(LT_B5, by="streets", mean=T)
}
}
