% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EmissionFactors.R
\name{EmissionFactors}
\alias{EmissionFactors}
\alias{print.EmissionFactors}
\title{Print for EmissionFactors}
\usage{
EmissionFactors(ef, ...)

\method{print}{EmissionFactors}(ef, all = TRUE)
}
\arguments{
\item{ef}{Object with class "EmissionFactors"}

\item{...}{ignored}

\item{all}{when T call method print.data.frame \code{\link{print.data.frame}}.
When F applies a  default print to each column}
}
\value{
Print method
}
\description{
A simple method for printing objects with class
"EmissionFactors".
}
\details{
Printing class "EmissionFactors"
}
\examples{
\dontrun{
# Do not run
data(fe2015)
names(fe2015)
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- as.EmissionFactors(df)
print(ef1)
print(ef1, all = F)
}
}
