% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Emissions.R
\name{print.Emissions}
\alias{print.Emissions}
\title{Printing class "Emissions"}
\usage{
print.Emissions(e, all = TRUE, ...)
}
\arguments{
\item{e}{Object with class "Emissions"}

\item{all}{when T call method print.data.frame \code{\link{print.data.frame}}.
When F applies a  default print to each column}
}
\value{
Print method
}
\description{
A simple method for printing objects with class "Emissions".
}
\examples{
\dontrun{
# Do not run
eco <- 2:15
as.Emissions(eco)
dfco <- as.Emissions(as.data.frame(eco))
print(dfco)
print(dfco, all = F)
}
}
