% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_units.R
\name{remove_units}
\alias{remove_units}
\title{Remove units}
\usage{
remove_units(x)
}
\arguments{
\item{x}{Object with class "data.frame", "matrix" or "units"}
}
\value{
"data.frame", "matrix" or numeric
}
\description{
\code{\link{remove_units}} Remove units from data.frames, matrix or units.
}
\examples{
{
ef1 <- ef_cetesb(p = "CO", c("PC_G", "PC_FE"))
class(ef1)
ef1
sapply(ef1, class)
a <- remove_units(ef1)
a
}
}
\keyword{units}
