% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_units.R
\name{remove_units}
\alias{remove_units}
\title{Remove units}
\usage{
remove_units(x, verbose = FALSE)
}
\arguments{
\item{x}{Object with class "sf",  "data.frame", "matrix" or "units"}

\item{verbose}{Logical, to print more information}
}
\value{
"sf", data.frame", "matrix" or numeric
}
\description{
\code{\link{remove_units}} Remove units from sf,  data.frames, matrix or units.
}
\examples{
\dontrun{
ef1 <- ef_cetesb(p = "CO", c("PC_G", "PC_FE"))
class(ef1)
sapply(ef1, class)
(a <- remove_units(ef1))
}
}
\keyword{units}
