% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_explode.R
\name{st_explode}
\alias{st_explode}
\title{Split line by vertex (experimental)}
\usage{
st_explode(net)
}
\arguments{
\item{net}{A spatial dataframe of class "sp" or "sf". When class is "sp"
it is transformed to "sf".}
}
\description{
\code{\link{st_explode}} split a lines data.frame into
each vertex. It to mimic the function explode from qgis, that the reason for the name
\code{\link{st_explode}}
}
\note{
All variables are transformed into numeric.

Thanks Michael Summer ([at]mdsummer)  for the function sfx_explode!
}
\examples{
{
data(net)
net <- sf::st_as_sf(net)
# transformign factor in numeric
net2 <- st_explode(net)
dim(net)
dim(net2)
}
}
