% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.EmissionFactors.R
\name{EmissionsFactors}
\alias{EmissionsFactors}
\alias{summary.EmissionFactors}
\title{Summary for EmissionFactors}
\usage{
EmissionsFactors(ef, ...)

summary.EmissionFactors(ef)
}
\arguments{
\item{ef}{Object of class EmissionFactors}

\item{...}{ignored}
}
\value{
Summary for vehicle classes
}
\description{
The object with class "EmissionFactors" is esentially a
data.frame with columns as type of vehicles and rows as age. The summary
method considers this by arguments by, mean and default to produce summaries
the information of EmissionFactors in a helpful way.
}
\details{
Summary of EmissionFactors
}
\examples{
\dontrun{
data(fe2015)
names(fe2015)
class(fe2015)
df <- fe2015[fe2015$Pollutant=="CO", c(ncol(fe2015)-1,ncol(fe2015))]
ef1 <- as.EmissionFactors(df)
summary(ef1)
}
}
\seealso{
\code{\link{summary}}
}
