% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Vehicles.R
\name{Vehicles}
\alias{Vehicles}
\alias{summary.Vehicles}
\title{Summary of Vehicles}
\usage{
Vehicles(veh, ...)

summary.Vehicles(veh, by = "col")
}
\arguments{
\item{...}{ignored}

\item{by}{This parameter allows 4 values: "col", "streets", "all" and
"default" "col" returns a summary of the sum of the columns
via \code{\link{colSums}}, "streets" summary of the sum of the streets
via \code{\link{rowSums}}, "all" returns a simple summary to each combunation
columns x street and "default" the default summary.data.frame via
\code{\link{summary.data.frame}}}
}
\value{
Summary for vehicle classes
}
\description{
The object with class "Vehicles" is esentially a data.frame
with columns as type of vehicles and rows as streets. The summary method
considers this by arguments by, mean and default to produce summaries
the information of vehicles in a helpful way.
}
\examples{
\dontrun{
data(net)
lt <- as.Vehicles(net$hdv)
class(lt)
plot(lt)
#with data.frames
LT_B5 <- age_hdv(x = lt,name = "LT_B5")
summary(LT_B5, by="age")
summary(LT_B5, by="streets")
summary(LT_B5, by="all")
summary(LT_B5, by="default")
}
}
\seealso{
\code{\link{summary}}
}
