% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.R
\name{to_latex}
\alias{to_latex}
\title{creates a .tex a table from a data.frame}
\usage{
to_latex(df, file, caption = "My table", label = "tab:df")
}
\arguments{
\item{df}{data.frame with three column.}

\item{file}{Character, name of new .tex file}

\item{caption}{Character caption of table}

\item{label}{Character,  label of table}
}
\value{
a text file with extension .tex.
}
\description{
\code{\link{to_latex}} reads a data.frme and generates a .tex
table, aiming to replicate the method of tablegenerator.com
}
\examples{
\dontrun{
ef <- ef_cetesb(p = "CO", veh = "PC_FG", full = T)
to_latex(ef)
}
}
\seealso{
\code{\link{vein_notes}}

Other helpers: 
\code{\link{colplot}()},
\code{\link{dmonth}()}
}
\concept{helpers}
