% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.R
\name{to_latex}
\alias{to_latex}
\title{creates a .tex a table from a data.frame}
\usage{
to_latex(df, file, caption = "My table", label = "tab:df")
}
\arguments{
\item{df}{data.frame with three column.}

\item{file}{Character, name of new .tex file}

\item{caption}{Character caption of table}

\item{label}{Character,  label of table}
}
\value{
a text file with extension .tex.
}
\description{
\code{\link{to_latex}} reads a data.frme an dgenerates a .tex
table, aiming to replicate the method of tablegenerator.com
}
\examples{
\dontrun{
df <- data.frame(pollutant = rep(c("CO", "propadiene", "NO2"), 10),
                 emission = vein::Emissions(1:30),
                 region = rep(letters[1:2], 15))
df
long_to_wide(df)
(df2 <- long_to_wide(df, column_fixed = "region"))
to_latex(df2)
to_latex(long_to_wide(df, column_fixed = "region"),
file = paste0(tempfile(), ".tex"))
}
}
\seealso{
\code{\link{vein_notes}}  \code{\link{long_to_wide}}

Other helpers: 
\code{\link{colplot}()},
\code{\link{dmonth}()},
\code{\link{wide_to_long}()}
}
\concept{helpers}
