% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.R
\name{to_latex}
\alias{to_latex}
\title{creates a .tex a table from a data.frame}
\usage{
to_latex(df, file = "dataframe.tex", caption = "My table",
  label = "tab:df", order = paste0("l", rep("r", ncol(df) - 1)))
}
\arguments{
\item{df}{data.frame with three column.}

\item{file}{Character, name of new .tex file}

\item{caption}{Character caption of table}

\item{label}{Character,  label of table}

\item{order}{Character,  alignment of columns of table, e.g.: "lrr"}
}
\value{
a text file with extension .tex.
}
\description{
\\code{\\link{to_latex}} reads a data.frme an dgenerates a .tex
table, aiming to replicate the method of tablegenerator.com
}
\examples{
{
df <- data.frame(pollutant = rep(c("CO", "propadiene", "NO2"), 10),
emission = vein::Emissions(1:30),
region = rep(letters[1:2], 15))
df
long_to_wide(df)
long_to_wide(df, column_fixed = "region")
to_latex(long_to_wide(df, column_fixed = "region"))
}
}
\seealso{
\\code{\\link{vein_notes}}  \\code{\\link{long_to_wide}}
}
