% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vehicles.R
\name{Vehicles}
\alias{Vehicles}
\title{Generic function of class "Vehicles"}
\usage{
Vehicles(veh, ...)
}
\value{
Construtor of objects with class "Vehicles"
}
\description{
Returns object with class "Vehicles"
}
