% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vehicles.R
\name{Vehicles}
\alias{Vehicles}
\alias{print.Vehicles}
\alias{summary.Vehicles}
\alias{plot.Vehicles}
\alias{print.Vehicles}
\alias{summary.Vehicles}
\alias{plot.Vehicles}
\title{Construction function for class "Vehicles"}
\usage{
Vehicles(x, ...)

\method{print}{Vehicles}(x, ...)

\method{summary}{Vehicles}(object, ...)

\method{plot}{Vehicles}(x, ...)
}
\arguments{
\item{x}{Object with class "Vehicles"}

\item{...}{ignored}

\item{object}{Object with class "Vehicles"}
}
\value{
Objects of class "Vehicles" or "units"
}
\description{
Returns a tranformed object with class "Vehicles" and units
 1/h. The type of objects supported are of classes "matrix", "data.frame",
 "numeric" and "array". If the object is a matrix it is converted to data.frame.
 If the object is "numeric" it is converted to class "units". The function
 \code{\link{emis_paved}} needs veh to be an array, therefore in this case,
 veh must be an array in the total fleet at each street and dimensions
 total fleet, hours and days
}
\examples{
\dontrun{
lt <- rnorm(100, 300, 10)
class(lt)
vlt <- Vehicles(lt)
class(vlt)
plot(vlt)
LT_B5 <- age_hdv(x = lt,name = "LT_B5")
print(LT_B5)
summary(LT_B5)
plot(LT_B5)
}
}
