% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vehicles.R
\name{Vehicles}
\alias{Vehicles}
\alias{print.Vehicles}
\alias{summary.Vehicles}
\alias{plot.Vehicles}
\title{Construction function for class "Vehicles"}
\usage{
Vehicles(x, ...)

\method{print}{Vehicles}(x, ...)

\method{summary}{Vehicles}(object, ...)

\method{plot}{Vehicles}(x, ..., message = TRUE)
}
\arguments{
\item{x}{Object with class "Vehicles"}

\item{...}{ignored}

\item{object}{Object with class "Vehicles"}

\item{message}{message with average age}
}
\value{
Objects of class "Vehicles" or "units"
}
\description{
\code{Vehicles} returns a tranformed object with class "Vehicles" and units
 'veh'. The type of objects supported are of classes "matrix", "data.frame",
 "numeric" and "array". If the object is a matrix it is converted to data.frame.
 If the object is "numeric" it is converted to class "units".
}
\examples{
{
lt <- rnorm(100, 300, 10)
class(lt)
vlt <- Vehicles(lt)
class(vlt)
plot(vlt)
LT_B5 <- age_hdv(x = lt,name = "LT_B5")
summary(LT_B5)
plot(LT_B5)
}
}
