% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsList.R, R/Evaporative.R, R/hot_soak.R,
%   R/matvect.R, R/running_losses.R, R/vein-deprecated.R
\name{vein-deprecated}
\alias{vein-deprecated}
\alias{EmissionsList}
\alias{print.EmissionsList}
\alias{summary.EmissionsList}
\alias{plot.EmissionsList}
\alias{Evaporative}
\alias{print.Evaporative}
\alias{summary.Evaporative}
\alias{plot.Evaporative}
\alias{hot_soak}
\alias{matvect}
\alias{running_losses}
\title{DEPRECATED Construction function for class "EmissionsList"}
\usage{
EmissionsList(x, ...)

Evaporative(x, ...)

hot_soak(x, carb, p, eshotc, eswarmc, eshotfi)

matvect(df, x, by = "row")

running_losses(x, erhotfi, lkm, carb, p, erhotc, erwarmc)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{ignored}

\item{df}{Numeric Data-frame or matrix.}

\item{by}{Character, with two value "row" or "col"}

\item{object}{Object with class "Evaporative"}
}
\value{
Objects of class "EmissionsList" and numeric elements as "units"

Objects of class "Evaporative" or "units"

data-frame
}
\description{
\code{EmissionsList} returns a tranformed object with class "EmissionsList".

\code{Evaporative} returns a tranformed object with class "Evaporative" and
units g. This class represents the daily emissions presented by
Mellios G and Ntziachristos (2016) Gasoline evaporation, Tier 2. Eventually
it will be incorporated the techniques of Tier 3.

\code{matvect} it is a helper function to multiply matrices
with vector by rows or columns

The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\section{\code{Evaporative}}{

For \code{Evaporative}, use \code{\link{emis_evap}}.


For \code{running_losses}, use \code{\link{emis_evap}}.


For \code{running_losses}, use \code{\link{emis_evap}}.
}

\examples{
{
# do not run
# DEPRECATED
}
{
# do not run
# DEPRECATED
}
{
# Do not run
}
{
# do not run
# DEPRECATED
}
{
# Do not run
}

}
\references{
Mellios G and Ntziachristos 2016. Gasoline evaporation. In:
EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European
Environment Agency, Copenhagen, 2009

Mellios G and Ntziachristos 2016. Gasoline evaporation. In:
EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European
Environment Agency, Copenhagen, 2009
}
\seealso{
\code{\link{vein-deprecated}}

\code{\link{vein-deprecated}}

\code{\link{vein-deprecated}}
}
\keyword{internal}
