% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein-package.R
\docType{package}
\name{vein-package}
\alias{vein}
\alias{vein-package}
\title{vein: Vehicular Emissions Inventories}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Elaboration of vehicular emissions inventories,
    consisting in four stages, pre-processing activity data, preparing 
    emissions factors, estimating the emissions and post-processing of emissions 
    in maps and databases. More details in Ibarra-Espinosa et al (2018) <doi:10.5194/gmd-11-2209-2018>.
    Before using VEIN you need to know the vehicular composition of your study area, in other words,
    the combination of of type of vehicles, size and fuel of the fleet. Then, it is recommended to
    start with the function inventory to create a structure of directories and template scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atmoschem.github.io/vein/}
  \item Report bugs at \url{https://github.com/atmoschem/vein/issues/}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (0000-0002-3162-1905)

Other contributors:
\itemize{
  \item Daniel Schuch \email{schuch@usp.br} (0000-0001-5977-4519) [contributor]
}

}
\keyword{internal}
