% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein_notes.R
\name{vein_notes}
\alias{vein_notes}
\title{vein_notes for writting technical notes about the inventory}
\usage{
vein_notes(file = "README", title, yourname, notes, overwrite = FALSE)
}
\arguments{
\item{file}{Character; Name of the file. The function will generate a file
with an extension '.txt'.}

\item{title}{Character; Title of this file. For instance: "Vehicular Emissions
Inventory of Region XX, Base year XX"}

\item{yourname}{Character; Name of the inventor compiler.}

\item{notes}{Character; vector of notes.}

\item{overwrite}{Logical. Do you want to overwrite existing file?}
}
\value{
A text file.
}
\description{
\code{\link{vein_notes}} creates aa text file '.txt' for
writting technical notes about this emissions inventory
}
\examples{
{
a <- tempfile()
vein_notes(a)
}
}
